move  export_mask() from the mixin to the frontend algorithm
