+++ Framework/PythonInterface/mantid/utils/nomad/diagnostics.py
@@ -13,7 +13,6 @@ import yaml
 # standard imports
 from collections import namedtuple
 import enum
-import os
 import random
 import string
 from typing import List, NamedTuple
@@ -300,7 +299,7 @@ class _NOMADMedianDetectorTest:
         -------
         1D array of size number-of-pixels
         """
-        from mantid.simpleapi import DeleteWorkspaces, Divide, SolidAngle
+        from mantid.simpleapi import DeleteWorkspaces, Divide, mtd, SolidAngle
         intensities_workspace = input_workspace  # clean up temporary workspaces
         if solid_angle_normalize:
             solid_angles, normalized_workspace = self._random_string(), self._random_string()  # temporary
@@ -458,42 +457,3 @@ class _NOMADMedianDetectorTest:
     @property
     def mask_by_pixel_intensity(self):
         return self.determine_tubes_threshold(self.intensities, self.config, self.set_nomad_constants())
-
-    @classmethod
-    def export_mask(cls,
-                    pixel_mask_states: numpy.ndarray,
-                    mask_file_name: str,
-                    instrument_name: str = 'NOMAD') -> None:
-        """
-        Export masks to XML file format
-
-        Parameters
-        ----------
-        pixel_mask_states: numpy.ndarray
-            boolean array with the number of pixels of NOMAD.  True for masking
-        mask_file_name: str
-            name of the output mask XML file or single-colun ASCII file
-        instrument_name: str
-            name of the instrument
-        """
-        from mantid.simpleapi import DeleteWorkspaces, ExtractMask, LoadEmptyInstrument, MaskDetectors, mtd, SaveMask
-        detector_ids_masked = np.where(pixel_mask_states)[0]  # detector ID's start at zero (monitors have negative ID)
-        if '.xml' in mask_file_name:
-            # Load empty instrument
-            empty_workspace_name = cls._random_string()
-            LoadEmptyInstrument(InstrumentName=instrument_name, OutputWorkspace=empty_workspace_name)
-
-            # Get the workspace indexes to mask. Shift by the number of monitors
-            if mtd[empty_workspace_name].getNumberHistograms() != pixel_mask_states.shape[0] + cls.MONITOR_COUNT:
-                raise RuntimeError(f'Spectra number of {instrument_name} workspace does not match mask state array')
-            mask_ws_indexes = detector_ids_masked + cls.MONITOR_COUNT
-
-            MaskDetectors(Workspace=empty_workspace_name, WorkspaceIndexList=mask_ws_indexes)
-
-            mask_workspace_name = cls._random_string()
-            ExtractMask(InputWorkspace=empty_workspace_name, OutputWorkspace=mask_workspace_name)
-            SaveMask(InputWorkspace=mask_workspace_name, OutputFile=mask_file_name)
-
-            DeleteWorkspaces([empty_workspace_name, mask_workspace_name])
-        else:
-            np.savetxt(mask_file_name, detector_ids_masked, fmt='%6d', newline=os.linesep)
